pyproject.toml
[project]
name = "lesekobold"
version = "1.1.0"
description = ""
authors = [
    {name = "zarah.weiss",email = "zarah.weiss@nextory.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"

dependencies = [
    "pydantic (>=2.12.4,<3.0.0)",
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "google-adk (>=1.18.0,<2.0.0)",
    "litellm (>=1.79.3,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]

[tool.poetry]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
ruff = "^0.14.5"


[tool.pytest.ini_options]
markers = [
    "unit_test: marks tests as proper unit tests",
    "makes_api_call: marks tests that make actual API calls without mocking them",
    "quality_test: marks tests used for manual quality checks",
    "performance_test: marks a test as a performance test (runtime or correctness)",
    "slow: marks tests that are slow to execute (deselected in CI/CD expect production)",
]
testpaths = ["tests"]
addopts = "-v -ra -q"
log_cli = true
log_cli_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
norecursedirs = "tests/test_agent"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
